Rename file and class after removal of tracking pixel
